youtube-bot/requirements.txt
discord.py
yt-dlp
PyNaCl
davey
python-dotenv
certifi
aiohttp==3.9.5
